// src/app/models/dnc-quotation/IDesignAndConsultancyModel.tsx
export interface IDNCQuotationModel {
  quotationID: number
  quotationNumber: string
  projectName: string
  projectNumber: string
  customerID: number
  quotationDate: string
  quotationAmount: string
  employeeID: number
  bhkName: string
  status: number
  bhkid: number
  carpetArea: string
  customerName: string
  email: string
  employeeName: string
  mobileNumber: string
  customerBranchName: string
  employeeBranchName: string
  quoteStatus: string
  isExtraDiscount: boolean
  extraDiscount: string
  reqExtraDisc: string
  discountStatusID: number
  discountCondition: string
  discountBy: number
  discountResponseBy: number
  discountTypeID: number
  extraDiscStatusName: string
  isBooked: boolean
  isBeforeDiscount: boolean
}
export const DNCQuotationInitValue: IDNCQuotationModel = {
  quotationID: 0,
  quotationNumber: '',
  projectName: '',
  projectNumber: '',
  customerID: 0,
  quotationDate: '',
  quotationAmount: '',
  employeeID: 0,
  bhkName: '',
  status: 0,
  bhkid: 0,
  carpetArea: '',
  customerName: '',
  email: '',
  employeeName: '',   
  mobileNumber: '',
  customerBranchName: '',
  employeeBranchName: '',
  quoteStatus: '',
  isExtraDiscount: false,
  extraDiscount: '',
  reqExtraDisc: '',
  discountStatusID: 0,
  discountCondition: '',
  discountBy: 0,
  discountResponseBy: 0,
  discountTypeID: 0,
  extraDiscStatusName: '',
  isBooked: false,
  isBeforeDiscount: false,
}